+++ packages/esm-patient-chart-app/src/dashboard.meta.ts
@@ -9,16 +9,3 @@ export const encountersDashboardMeta = {
   path: 'Visits',
   title: 'Visits',
 };
-
-//clinical views
-export const hivPatientSummaryDashboardMeta = {
-  slot: 'hiv-care-and-treatment-group-slot',
-  path: 'hiv-patient-summary',
-  title: 'HIV Patient Summary',
-};
-
-export const programManagementDashboardMeta = {
-  slot: 'hiv-care-and-treatment-group-slot',
-  path: 'hiv-program-management',
-  title: 'HIV Program Management',
-};
