+++ packages/esm-patient-chart-app/src/encounter-tile/components/tile.scss
@@ -3,7 +3,7 @@
 @use '@openmrs/esm-styleguide/src/vars' as *;
 
 .title {
-  @include type.productiveHeading03;
+  @include type.type-style('productive-heading-03');
 }
 
 .title:after {
@@ -21,7 +21,7 @@
 }
 
 .tile {
-  height: 100%;
+  height: 35%;
   padding: layout.$spacing-01 0 layout.$spacing-05 layout.$spacing-05;
   margin: layout.$spacing-03;
   border: solid 1px #e0e0e0;
