+++ packages/esm-patient-chart-app/src/encounter-tile/utils/helpers.ts
