+++ packages/esm-patient-chart-app/src/index.ts
@@ -12,12 +12,7 @@ import genericDashboardComponent, { genericDashboardConfigSchema } from './side-
 import genericNavGroupComponent, { genericNavGroupConfigSchema } from './side-nav/generic-nav-group.component';
 import { moduleName } from './constants';
 import { setupCacheableRoutes, setupOfflineVisitsSync } from './offline';
-import {
-  summaryDashboardMeta,
-  encountersDashboardMeta,
-  hivPatientSummaryDashboardMeta,
-  programManagementDashboardMeta,
-} from './dashboard.meta';
+import { summaryDashboardMeta, encountersDashboardMeta } from './dashboard.meta';
 import addPastVisitActionButtonComponent from './actions-buttons/add-past-visit.component';
 import cancelVisitActionButtonComponent from './actions-buttons/cancel-visit.component';
 import currentVisitSummaryComponent from './visit/visits-widget/current-visit-summary.component';
@@ -127,22 +122,6 @@ export const encountersSummaryDashboardLink =
     }),
     { featureName: 'encounter', moduleName },
   );
-
-export const hivPatientSummaryDashboardLink = getSyncLifecycle(
-  createDashboardLink({
-    ...hivPatientSummaryDashboardMeta,
-    moduleName,
-  }),
-  { featureName: 'hiv-patient-summary', moduleName },
-);
-
-export const hivPatientManagementDashboardLink = getSyncLifecycle(
-  createDashboardLink({
-    ...programManagementDashboardMeta,
-    moduleName,
-  }),
-  { featureName: 'hiv-patient-summary', moduleName },
-);
 
 export const currentVisitSummary = getSyncLifecycle(currentVisitSummaryComponent, {
   featureName: 'current-visit-summary',
