--- packages/esm-patient-chart-app/src/patient-summary/program-management.component.tsx
@@ -1,6 +1,0 @@
-import React from 'react';
-import { EmptyState } from '@openmrs/esm-patient-common-lib';
-
-export function ProgramManagement() {
-  return <EmptyState headerTitle="Program Management" displayText="Coming soon!" />;
-}
