--- packages/esm-patient-chart-app/src/patient-summary/service-summary-config.json
@@ -1,137 +1,0 @@
-{
-  "encounter-tiles-group" : {
-    "tilesDefinitions" : [
-      {
-        "title": "Characteristics",
-        "columns": [
-          {
-            "id": "artCohort",
-            "title": "ART Cohort",
-            "encounterType": "art_Therapy_EncounterUUID",
-            "concept": "artTherapyDateTime_UUID",
-            "isDate": true,
-            "conceptMappings": [
-              "artTherapyDateTime_UUID",
-              "artStopDateUUID",
-              "switchDateUUID",
-              "substitutionDateUUID",
-              "dateRestartedUUID"
-            ]
-          },
-          {
-            "id": "currentRegimen",
-            "hasSummary": true,
-            "title": "Current Regimen",
-            "encounterType": "art_Therapy_EncounterUUID",
-            "concept": "regimenConcept",
-            "summaryConcept": {
-              "primaryConcept": "regimenLine_UUID"
-            }
-          },
-          {
-            "id": "dsdModel",
-            "title": "DSD Model",
-            "hasSummary": true,
-            "encounterType": "ServiceDeliveryEncounterType_UUID",
-            "concept": "CommunityDSDModel_UUID",
-            "summaryConcept": {
-              "primaryConcept": "CommunityDSDModel_UUID"
-            }
-          },
-          {
-            "id": "populationType",
-            "title": "Population Type",
-            "encounterType": "careAndTreatmentEncounterType",
-            "concept": "populationCategoryConcept",
-            "hasSummary": true,
-            "summaryConcept": {
-              "primaryConcept": "keyPopulationTypeConcept",
-              "secondaryConcept": "priorityPopulationTypeConcept"
-            }
-          }
-        ]
-      },
-      {
-        "title": "HIV Monitoring",
-        "columns": [
-          {
-            "id": "viralLoad",
-            "title": "Current Viral Load",
-            "encounterType": "ViralLoadResultsEncounter_UUID",
-            "concept": "ViralLoadResult_UUID",
-            "hasSummary": true,
-            "summaryConcept": {
-              "primaryConcept": "ViralLoadResultDate_UUID",
-              "isDate": true
-            }
-          },
-          {
-            "id": "currentVLReason",
-            "title": "Reason For Current VL",
-            "encounterType": "art_Therapy_EncounterUUID",
-            "concept": "ReasonForViralLoad_UUID",
-            "hasSummary": true,
-            "summaryConcept": {
-              "primaryConcept": "ViralLoadResultDate_UUID",
-              "isDate": true,
-              "hasCalculatedDate": true
-            }
-          },
-          {
-            "id": "lastCD4Count",
-            "title": "Last CD4 Count",
-            "hasSummary": true,
-            "encounterType": "CD4LabResultsEncounter_UUID",
-            "concept": "hivCD4Count_UUID",
-            "summaryConcept": {
-              "primaryConcept": "Cd4LabResultDate_UUID",
-              "isDate": true
-            }
-          }
-        ]
-      },
-      {
-        "title": "Last Visit Details",
-        "columns": [
-          {
-            "id": "tbScreening",
-            "title": "TB Screening",
-            "encounterType": "clinicalVisitEncounterType",
-            "concept": "tbScreeningOutcome",
-            "hasSummary": true,
-            "summaryConcept": {
-              "primaryConcept": "regimenLine_UUID",
-              "isDate": true
-            }
-          },
-          {
-            "id": "oIs",
-            "title": "OIs",
-            "encounterType": "clinicalVisitEncounterType",
-            "concept": "opportunisticInfectionConcept"
-          },
-          {
-            "id": "nextAppointmentDate",
-            "title": "Next Appointment Date",
-            "hasSummary": true,
-            "encounterType": "clinicalVisitEncounterType",
-            "concept": "returnVisitDateConcept",
-            "summaryConcept": {
-              "primaryConcept": "returnVisitDateConcept",
-              "isDate": true,
-              "hasCalculatedDate": true
-            }
-          },
-          {
-            "id": "programStatus",
-            "title": "Program Status",
-            "encounterType": "hivProgramStatusEncounterType",
-            "concept": "generalTreatmentStatusConcept"
-          }
-        ]
-      }            
-    ]
-  }
-}
-
-
